calcula_dv_municipio

+++ packages/cms/src/Controllers/SerieController.php
@@ -772,8 +772,10 @@ class SerieController extends Controller
         $peso = "1212120";
         //echo substr($peso,0,1);
         $soma = 0;
-        for($i = 0; $i < 7; $i++){ $valor = substr($codigo_municipio,$i,1) * substr($peso,$i,1); if($valor>9)
-            $soma = $soma + substr($valor,0,1) + substr($valor,1,1);
+        for($i = 0; $i < 7; $i++){
+
+            $valor = (int)substr($codigo_municipio,$i,1) * (int)substr($peso,$i,1); if($valor>9)
+            $soma = (int)$soma + (int)substr($valor,0,1) + (int)substr($valor,1,1);
         else
             $soma = $soma + $valor;
         }
